Broken commit - Prep before webextension-polyfill integration
So. I tried to implement some of the storage setting and getting stuff and got in to a mess. Many wasted hours later, I stumbled over webextension-polyfill (already in project from template!) and this commit is to draw a line under things before I integrate it. Like a boss.

Damn this is going to make stuff easier.